Clean up TestRunSpecDetails by refactoring part of it into TestRunResultList
